+++ config/cloud.env
@@ -3,7 +3,5 @@ SUBDOMAIN=officetracker
 DOMAIN=baileys.page
 BASE_PATH=/
 GH_CLIENT_ID=927f8338aecb75c09854
-NTFY_TOPIC=xbd_au
-BACKEND_ENDPOINT=https://officetracker.baileys.page/form
 PROJECT_ID=baileybutler-syd
 COLLECTION_ID=officetracker
