+++ internal/database/db.go
@@ -1,174 +1,20 @@
 package database
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"time"
-
-	"cloud.google.com/go/firestore"
-	"google.golang.org/api/iterator"
+	"github.com/baely/officetracker/internal/models"
 )
 
-type Client struct {
-	*firestore.Client
+type Databaser interface {
+	SaveEntry(e models.Entry) error
+	GetEntries(userID string, month, year int) (models.Entry, error)
+	GetAllEntries(userID string) ([]models.Entry, error)
+	GetEntriesForBankYear(userID string, bankYear int) ([]models.Entry, error)
 }
 
-type Entry struct {
-	User        string
-	CreateDate  time.Time
-	Month, Year int
-	Days        map[string]int
-	Notes       string
-}
-
-func buildDocumentId(e Entry) string {
-	return fmt.Sprintf("%s-%d-%d", e.User, e.Month, e.Year)
-}
-
-func generateDocumentId(userID string, day, month, year int) string {
-	return fmt.Sprintf("%s-%d-%d-%d", userID, day, month, year)
-}
-
-func NewFirestoreClient() (*Client, error) {
-	ctx := context.Background()
-	projectID := os.Getenv("PROJECT_ID")
-	client, err := firestore.NewClient(ctx, projectID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create firestore client: %v", err)
-	}
-	return &Client{client}, nil
-}
-
-func (c *Client) SaveEntry(e Entry) (string, error) {
-	ctx := context.Background()
-	collection := os.Getenv("COLLECTION_ID")
-	docId := buildDocumentId(e)
-	_, err := c.Collection(collection).Doc(docId).Set(ctx, e)
-	if err != nil {
-		return "", fmt.Errorf("failed to save entry: %v", err)
-	}
-	return docId, nil
-}
-
-func (c *Client) GetEntries(userId string, month, year int) (Entry, error) {
-	ctx := context.Background()
-	collection := os.Getenv("COLLECTION_ID")
-	docTitle := buildDocumentId(Entry{
-		User:  userId,
-		Month: month,
-		Year:  year,
-	})
-	doc, err := c.Collection(collection).Doc(docTitle).Get(ctx)
-	if !doc.Exists() {
-		return Entry{}, nil
-	}
-	if err != nil {
-		return Entry{}, fmt.Errorf("failed to fetch entry: %v", err)
+func NewClient(standalone bool) (Databaser, error) {
+	if standalone {
+		return newSQLiteClient()
 	}
 
-	var e Entry
-	if err := doc.DataTo(&e); err != nil {
-		return Entry{}, fmt.Errorf("failed to fetch entry: %v", err)
-	}
-
-	return e, nil
-}
-
-func (c *Client) GetAllEntries(userId string) ([]Entry, error) {
-	ctx := context.Background()
-	collection := os.Getenv("COLLECTION_ID")
-	iter := c.Collection(collection).
-		Where("User", "==", userId).
-		OrderBy("Year", firestore.Asc).
-		OrderBy("Month", firestore.Asc).
-		OrderBy("Day", firestore.Asc).
-		Documents(ctx)
-	defer iter.Stop()
-
-	var entries []Entry
-	for {
-		doc, err := iter.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			return entries, fmt.Errorf("failed to iterate entries: %v", err)
-		}
-
-		var e Entry
-		if err := doc.DataTo(&e); err != nil {
-			return entries, fmt.Errorf("failed to convert entry: %v", err)
-		}
-		entries = append(entries, e)
-	}
-	return entries, nil
-}
-
-func (c *Client) GetLatestEntries(userId string) ([]Entry, error) {
-	allEntries, err := c.GetAllEntries(userId)
-	if err != nil {
-		return nil, err
-	}
-
-	var entries []Entry
-	for _, e := range allEntries {
-		entries = append(entries, e)
-	}
-
-	return entries, nil
-}
-
-func (c *Client) GetEntriesForBankYear(userID string, bankYear int) ([]Entry, error) {
-	ctx := context.Background()
-	collection := os.Getenv("COLLECTION_ID")
-	var entries []Entry
-
-	iterPrev := c.Collection(collection).
-		Where("User", "==", userID).
-		Where("Year", "==", bankYear-1).
-		Where("Month", ">=", 10).
-		OrderBy("Month", firestore.Asc).
-		Documents(ctx)
-
-	for {
-		doc, err := iterPrev.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			return entries, fmt.Errorf("failed to iterate entries: %v", err)
-		}
-
-		var e Entry
-		if err := doc.DataTo(&e); err != nil {
-			return entries, fmt.Errorf("failed to convert entry: %v", err)
-		}
-		entries = append(entries, e)
-	}
-
-	iter := c.Collection(collection).
-		Where("User", "==", userID).
-		Where("Year", "==", bankYear).
-		Where("Month", "<=", 9).
-		OrderBy("Month", firestore.Asc).
-		Documents(ctx)
-
-	for {
-		doc, err := iter.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			return entries, fmt.Errorf("failed to iterate entries: %v", err)
-		}
-
-		var e Entry
-		if err := doc.DataTo(&e); err != nil {
-			return entries, fmt.Errorf("failed to convert entry: %v", err)
-		}
-		entries = append(entries, e)
-	}
-
-	return entries, nil
+	return newFirestoreClient()
 }
